snmp-simple/requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
# --extra-index-url http://mirrors.aliyun.com/pypi/simple/

PyMySQL==0.7.11
requests==2.18.4
apscheduler==3.5.1

pysnmp==4.4.9
xmltodict==0.12.0

